Create intermediate sorted index chunk files
Inserting into sled and lmdb for unordered keys is too slow (thousands/sec), but inserting ordered keys into lmdb is very fast (millions/sec). So, we write a bunch of chunks, each of which is sorted internally. We can then merge-sort the whole group of them into one sorted keyspace that can be written quickly to lmdb.

Writing chunks is fast (10-20M/sec), so as long as merge sort is implemented reasonably quickly, the overall time should still be quite reasonable.